\name{TMAX}
\alias{TMAX}

\title{
Daily High Temperature data
}
\value{
    \item{\code{TMAX}}{ - a numeric matrix}
}
\description{
TMAX is a user provided matrix of daily high temperatures (deg F). The matrix should be
366 rows (366 days) by as many years as desired in columns. Example high temperature
data is available if one calls data("BatonRouge")
}
