\name{TMIN}
\alias{TMIN}

\title{
Daily Low Temperature data
}
\value{
    \item{\code{TMIN}}{ - a numeric matrix}
}
\description{
TMIN is a user provided matrix of daily low temperatures (deg F). The matrix should be
366 rows (366 days) by as many years as desired in columns. Example low temperature
data is available if one calls data("BatonRouge")
}


