\name{YEAR}
\alias{YEAR}

\title{
Years of the Data
}
\value{
    \item{\code{YEAR}}{ - a numeric vector}
}
\description{
YEAR is a user provided vector of the years in the TMAX and TMIN data used with calc_si.
YEAR is not used in the calc_si (or any internal functions), but it is used for
plotting the results in the example code.
Example of YEAR is available if one calls data("BatonRouge")
}

